Seperated out the logic from the Header component for the Search Component into it's own file and made sure it rendered
